tests: fix test-subrepos.t sporadically failing
I was seeing sporadic failures running this test on Mac OS X 10.8.
They looked like this:
+  sed: 1: "s_/private/var/folders/ ...": bad flag in substitute command: 'T'

My assumption is that some character was being included in the path of the
temporary directory that sed didn't like.  It looks like the sed function was
being used to eliminate differences between test runs due to the path changing
each run.  That isn't needed any more now that we're using the unified test
format, since said replacement is taken care of for us by run-tests.py.  Thus,
this changeset removes the calls to sed and updates the output to use the result
from the framework-level replacement.