context: update to work with upstream
Some upstream code now relies on ctx.repo() to access the repo object, since
hggit stored ctx.repo as a field instead of a function, that failed. Let's move
it to be a function.

For overlaychangectx we can't just store the repo in self._repo because that
causes other parts of the base class to act differently and causes tests to
fail, so we store it as a new field.

--HG--
extra : histedit_source : acaa45385ebe5ae72e9638a76dc6d7b8c3110e10