tests: fix bad (re) match
It seems fffe8883960b incorrectly did a match on the second
parent (which is null) and that core Mercurial mistakenly matched the
whole line with (re). This was recently fixed in 6d5718e39657 in core,
so let's update the test with the correct match.

--HG--
extra : rebase_source : a4617e193cce6065974f595b50e23a17ecb5939f