heptapod-ci: adding versions that pass the tests down to 4.3
Currently, hg-git does not pass the tests for

- Mercurial 4.6
- the head of Mercurial default branch

In order to start benefit of the CI and enforcing that contributions must
pass it, we will therefore add jobs for these two versions in separate
topics and Merge Requests.

Also worth of note is that 5.2 can have new point releases that will be
provided in the base Docker image. Hence we don't exclude it from nightly
runs.