heptapod-ci: fixed job for Mercurial 4.6 by silencing warnings
Without these two flags, we get the following:

```
  devel-warn: changectx.__init__ is getting more limited, see context.changectxdeprecwarn() for details
  (compatibility will be dropped after Mercurial-4.6, update your code.) at: /tmp/hgtests.ShCiwV/install/lib/python/mercurial/localrepo.py:849 (__contains__)
```

after inspection of the code paths, it turns out that these are self-inflicted
Mercurial warnings. This can be seen with test-clone.t, the whole traceback
does not involve hg-git at all.

So the only practical solution we have is to silence the warnings.
I think it's acceptable as long as it's for this single version and in
particulary not for any version under active development or maintenance.

--HG--
extra : source : 9e5d488a2d4b933c9b60816188b767c64d1afc5f
extra : topic : ci-4.6