py3: byteify sources
These files were byteified using hg’s contrib/byteify-strings.py:

- contrib/hggitperf.py
- hggit/compat.py
- hggit/git2hg.py
- hggit/gitdirstate.py
- hggit/gitrepo.py
- hggit/git_handler.py
- hggit/hg2git.py
- hggit/hgrepo.py
- hggit/overlay.py
- hggit/util.py
- hggit/verify.py
- hggit/_ssh.py
- hggit/__init__.py
- tests/commitextra.py

Additionally, "--treat-as-kwargs opts kwargs" were passed as arguments.

Some of the changes will be undone later. Because of the large amount of
literals that should be changed in comparison to the few literals that should
stay, it seemed better to do it this way around.

Raw literals are not changed by the script. Some of them will be changed
manually in a separate changeset.