push: don't create git branches without --new-branch or --bookmark
The push semantics differ quite a bit between Mercurial & Git:

* Mercurial defaults to pushing everything, but never pushes bookmarks
* Git defaults to pushing just the 'current' branch

Combined with the fact that Mercurial doesn't delete bookmarks on
pull, you have a situation where pushing to a Git repository easily
revives deleted branches/bookmarks. That's just annoying.

With this change, the semantics are the same when pushing bookmarks,
regardless of whether you're targeting Mercurial or Git. Since
bookmarks and branches are the same when pushing to Git, we overload
`--new-branch` to push all bookmarks.