ci: retry failing jobs
The compatibility tests occasionally fail. To work around this, we
simply set the tests to retry once. Given that each test run is quite
short, we can safely retry them. We do so only once to begin with, but
to be safe, I've also added a timeout so that a hanging job isn't left
running for half an hour.