tag command: made usable as a function
This makes the `git` keyword argument effectively optional in
the function backing the `tag` command, that some downstream
code may call directly, without being aware the argument exists.

An example would be py-heptapod tests, and more generally tests
based on mercurial-testhelpers that needs to work regardless
of if hg-git is activated or not.