ci: triggering the downstream py-heptapod project
The `strategy: depend` makes sure that if downstream pipeline fails,
the upstream job is marked as failed (but currently py-heptapod's hg-git
job allows failures).

Passing down the current Mercurial sha is essential for the downstream
pipeline to test the correct version. The presence of this variable
will also allow it to skip irrelevant jobs.


WIP ci: depending on status of downstream triggered py-heptapod