obsolescence: add a config option to trigger inferred obsmarker creation
This is the first stone of a new mechanism to automatically create obsolescence
markers to link local draft content with content pulled from the main
repository. This should "reconcile" action done by other git client on local
changesets.

The heuristic is currently very basic, it check that the diff are the same
(except for the line numbers). We currently focus on the ability to create
Obsolescence markers on pull, we can "easily" improve the heuristic later.