clone: always use a single transaction
Given that Mercurial will delete the repository on any error, using
anything more than a single transaction is a waste of resources. In
fact, the only way to “recover” a partial clone using hg-git is to
forcibly KILL the process. This makes us match the behaviour of
Mercurial: clone isn't resumable, period, and not even inspecting the
repository during the clone suggests otherwise.

This currently doesn't affect the current defaults, though, which are to
always use a single transaction.