pull: adjust default map save interval to 1000
Previously, users would have to set this manually to make long running
conversions safely resumable. Now that we disregard it for clone, we
should make our defaults safer for pull.

The value of 1000 is essentially a compromise, as it isn't so small
that it makes converting repositories with more commits than files
slow, and not so large that it becomes infrequent with large and slow
commits.

--HG--
extra : rebase_source : 8589657b0381fc305fc97f0f4d69e13b6846918c
extra : topic : optimise-clone