clone: move in pack when possible
Dulwich has the add_thin_pack() function for adding a pack to the
repository which may be thin, i.e. refer to objects outside the pack.
We can avoid that on an initial clone, however, as we know that we
just created the Git repository, meaning that the pack is never thin.

The overhead is negligible on smaller repositories, but can be
comparable to network transfer on a fast connection. The only downside
is that it just might break something if the user specifies a
preexisting Git repository in the configuration, but that seems a bit
esoteric.