core: transfer git commit map and tags between hg repos
This implements transferring the Git state between Mercurial
repositories, and saving it in bundles. The behaviour is disabled by
default, hidden behind two experimental options.

We do so by embedding optional bundle2 parts that contain tags and git
mapping for the changesets in the bundle2. For local clones, we add
the mapfile and tags file to list of files to copy.

There is no support for reinstating or synchronising lost or old
state, and no support for transferring tags to old changesets. The
main upside to this is that we've restricted the potential for
conflicts massively.

Since this makes it more likely that users will run into the case
where a git parent is missing, I've tweaked the error message.