incoming: don't export commits first
This seems somewhat counter-intuitive to me. Incoming should ideally
be a cheap operation that, at the cost of network overhead, doesn't do
anything.

My best guess is that this exists to handle the case where commits
exist in Mercurial, that someone else already pushed to Git. In that
case, one might conceivably argue that they are, in fact incoming from
Git.