tests: add diagnostic output to test-serve-dulwich.t
We have a regression in this file, that appears to be caused by a commit
to Dulwich,[1] that enabled chunked upload. This just adds a bit of
output that reveals the exact error.

https://github.com/jelmer/dulwich/commit/e9faf83ac437d59bea8ef82d14c87c147e4a508a

--HG--
branch : 1.0.x
extra : rebase_source : 8da902db01588fb39e7915e929c1cdad2eefbe6f
extra : topic : dulwich-diagnostics