handler: ignore progress lines that cannot be parsed
Previously, the code assumes every line that ends with `\r` is a progress line
that contains a number.  This assumption is broken by the response returned by
Azure Repos, which contains an empty line that ends with `\r` and causes the
extension to crash.

--HG--
branch : 1.0.x
extra : rebase_source : 1d7483a36c6fbb271dcb6c52567cce13bc38cb9d
extra : topic : handler-ignore-progress-lines-that-cannot-be-parsed