pull: only write unchanged refs
Pulling from a repository that contains a lot of tags and branches
(over 2000) is very slow even when there is nothing to do. E.g. the
following output shows up quickly:

  > hg pull
  pulling from git+https://redacted
  no changes found

Then I have to wait over 2 minutes for the command to complete and the
shell prompt to return. I was able to trace the issue to
git_handler.py:1933 and git_handler.py:1937 which get called for each
ref and trigger a call to fsync inside dulwich. I was able to fix the
issue by checking if we're really trying to set a new hash or if we're
just rewriting the same one.