hggit: fix backwards compatibility with extrainmessage
When hg-git-rename-source was introduced, it was added unconditionally
in order to not break bidirectionality, however, that only applies if
debugextrainmessage is not set, if it's set 'HG:rename-source' is not
added.

The hg-git-rename-source change broke compatibility with git-remote-hg,
the following patch restores it back.

--HG--
extra : rebase_source : 74dd813da1d72cdc365b84ebf86a7a5c4479575d
extra : topic : fix-hg-git-rename-source