core: remove debugextrainmessage support
The stated purpose for this is for testing. However, it's not for
testing of hg-git, but of git-remote-hg. The mode has been broken for
years, though, and given that it doesn't actually rely on the
known-good modes of hg-git, it doesn't offer much guarantee of
anything actually working.

Removing as general cruft that needlessly complicates the codebase.

Closes: #281 !211