core: remove debugextrainmessage support
The stated purpose for this is for testing. However, it's not for
testing of hg-git, but of git-remote-hg. The mode has been broken for
years, though, and given that it doesn't actually rely on the
known-good modes of hg-git, it doesn't offer much guarantee of
anything actually working.

Removing as general cruft that needlessly complicates the codebase.

Closes: #281 !211

--HG--
extra : rebase_source : 1fa5411c414e21259c9583fd20d724e75054b58d
extra : topic : kill-debugextrainmessage