gitrepo: provide the original opargs to exchange.push when repo is not gitrepo
We got a report in evolve/topic extension that hg push --publish was broken
when both topic and hggit was enabled. Disabling either of them made it work
again. When searching what could've went wrong, I noticed that opargs provided
to exchange.push() don't contain 'publish': True anymore, here's the reason.
This is a solution to the issue, it might not be enough in all situations, but
it seems to be the correct thing to do in this particular block of code.

This call of orig() is only when repo is not gitrepo, IOW it's only for regular
Mercurial repos.