util: don't attempt to delete an open NamedTemporaryFile (fixes #440)
On Windows, I had problems both cloning and pulling new changes from the server.
The workaround was the clone/pull in WSL using Linux python.  Not sure when this
broke, but apparently `tempfile.NamedTemporaryFile` has a lot of... limitations
on Windows.[1]

My suggestion of using `mercurial.util.unlink()` in the bug report didn't work
for pulling changes.  I didn't test if it would have worked for cloning.

Note that this isn't a problem on the default branch, because support for
`dulwich` prior to 0.20 drops this code.  It will generate a merge conflict, but
can be thrown away.

[1] https://stackoverflow.com/a/49868505